+++ infra/modules/azure_postgres_server/CHANGELOG.md
@@ -1,5 +1,23 @@
 # azure_postgres_server
 
+## 1.0.1
+
+### Patch Changes
+
+- 8c05dd7: Removed HA if WEU as Default, added override variable for HA
+
+## 1.0.0
+
+### Major Changes
+
+- bc3027b: L tier now uses GP_Standard_D4ds_v5 as SKU, and M tier uses GP_Standard_D2ds_v5
+
+## 0.1.1
+
+### Patch Changes
+
+- 16ecc30: Using a common resource group in terraform tests
+
 ## 0.1.0
 
 ### Minor Changes